Add files via upload
Small Basket is a huge online / mobile application-based grocery retailer in India, founded
in 2011. Small Basket is trying to manage its supply chain and delivery partners and would
like to accurately forecast the sales for the period starting from January of 2019 to August
of 2019. You are also given a few features that were developed by the Business
Intelligence team, that might or might not provide additional insights to your analysis.

In this hackathon, you are tasked with creating a model that can accurately forecast the
sales for small basket across several locations and product categories.